requirements.txt
fastapi==0.104.1
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
pillow>=10.0.0,<11.0.0
python-multipart==0.0.6
spacy>=3.7.0,<4.0.0
thinc>=8.1.0,<9.0.0
torch==2.1.0+cpu
torchvision==0.16.0+cpu
tqdm>=4.65.0
uvicorn[standard]>=0.24.0
azure-storage-blob>=12.27.0
--extra-index-url https://download.pytorch.org/whl/cpu
